requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
decorator==5.2.1
Django==5.2
django-apscheduler==0.7.0
executing==2.2.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.1.0
pytils==0.4.3
redis==5.2.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzlocal==5.3.1
wcwidth==0.2.13
